+++ workbench/libs/muimaster/classes/scale.c
@@ -177,7 +177,7 @@ IPTR Scale__MUIM_Draw(struct IClass *cl, Object *obj, struct MUIP_Draw *msg)
 	    {
 		//int pct = j * 100 / i;
 
-		snprintf(buf, 255, "%d%%", j);
+		snprintf(buf, 255, "%ld%%", (LONG)j);
 		total_width += TextLength(_rp(obj),buf,strlen(buf));
 		if (total_width > (3 * _mwidth(obj) / 8))
 		{
@@ -223,7 +223,7 @@ IPTR Scale__MUIM_Draw(struct IClass *cl, Object *obj, struct MUIP_Draw *msg)
 		ZText *ztext;
 		int val = k * 100 / i;
 
-		snprintf(buf, 255, "%d%%", val);
+		snprintf(buf, 255, "%ld%%", (LONG)val);
 		ztext = zune_text_new(NULL,buf,ZTEXT_ARG_NONE,0);
 		if (ztext)
 		{
